# fix(mv3-part-8): Only persist store data when changed